scholarship-finder/requirements.txt
# Web scraping libraries
requests==2.32.4
beautifulsoup4==4.12.3

# Data processing
pandas==2.2.2
numpy==1.26.4

# Database
pymysql==1.1.0



# Utilities
python-dotenv==1.0.1
lxml==5.2.1
html5lib==1.1
fake-useragent==1.4.0
tenacity==8.2.3
# Pin httpx to a version compatible with openai==1.12.0 (proxies kwarg supported)
httpx==0.27.2

# AI and API libraries
# Use a recent 1.x OpenAI SDK; keep httpx pinned below 0.28 due to proxies change
openai>=1.40.0,<2.0.0

# Development
pytest==8.2.2
black==24.3.0
flake8==7.1.1
ruff==0.6.9
